refactor: investment info page refactor to investment schedule page
