+++ src/entities/element/Button.tsx
@@ -51,15 +51,18 @@ const ButtonWrapper = styled.button<{ variant: string | undefined }>`
   }
 `;
 
-export const BlueBorderButton = styled.button<{ width?: string, height?: string }>`
-  border: 1px solid #0500FF;
+export const BlueBorderButton = styled.button<{
+  width?: string;
+  height?: string;
+}>`
+  border: 1px solid #0500ff;
   border-radius: 8px;
   width: ${(props) => props.width || "100%"};
   height: ${(props) => props.height || "20px"};
   padding: 10px 0px;
   display: flex;
   justify-content: center;
-  color: #0500FF;
+  color: #0500ff;
   cursor: pointer;
 `;
 export const SignButton = styled.button`
@@ -71,6 +74,6 @@ export const SignButton = styled.button`
   width: 100%;
   font-size: 16px;
   color: white;
-  background-color: #0500FF;
+  background-color: #0500ff;
   cursor: pointer;
 `;
